sched/tune: Rename dynamic_boost parameter to sched_boost
This was confusing to deal with given that it had the same name as the
Dynamic Schedtune Boost framework. It will be more apt to call it
sched_boost given that it was created to work with the sched_boost
feature in CAF devices.

The new tunable can be found in /dev/stune/*/schedtune.sched_boost

Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: Henrique Pereira <hlcpereira@pixelexperience.org>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>
Signed-off-by: madeofgreat <ravenklawasd@gmail.com>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>